services/lambda_function/requirements.txt
pyodbc
boto3
psycopg2-binary
